fix: use portable format specifiers for int64_t cross-platform compatibility
- Added <inttypes.h> header to modules formatting int64_t values
- Replaced all %lld format specifiers with portable PRId64 macro
- Updated 6 modules for consistent portable formatting
- Removed (long long) casts in favor of type-safe macros
- Fixed Linux compilation errors where int64_t is long not long long
- Added portable format specifier guidelines to CLAUDE.md
- Created AGENTS.md with reference to follow CLAUDE.md guidelines

+++ modules/agerun_agent_registry.c
@@ -12,6 +12,7 @@
 #include <stdlib.h>
 #include <stdint.h>
 #include <string.h>
+#include <inttypes.h>
 
 /* Agent registry structure */
 struct agent_registry_s {
@@ -203,7 +204,7 @@ bool ar__agent_registry__register_id(agent_registry_t *mut_registry, int64_t age
     
     // Create string representation of the ID
     char id_str[32];
-    snprintf(id_str, sizeof(id_str), "%lld", (long long)agent_id);
+    snprintf(id_str, sizeof(id_str), "%" PRId64, agent_id);
     
     // Create string data for the ID (ar__data__create_string copies the string)
     data_t *own_id_data = ar__data__create_string(id_str);
